Final AirlineApplication

+++ src/main/java/com/moja/airline/entity/user/UserRepository.java
@@ -1,11 +1,10 @@
 package com.moja.airline.entity.user;
 
-import com.moja.airline.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-    public interface UserRepository  extends JpaRepository<User, Integer> {
+    public interface UserRepository  extends JpaRepository<User, Long> {
     public User findByUsername(String username);
 
     }
